+++ cmd/main.go
@@ -172,6 +172,14 @@ func main() {
 		os.Exit(1)
 	}
 
+	if err := (&controller.CustomInfoKeyReconciler{
+		Client:  mgr.GetClient(),
+		Clients: clientPool,
+	}).SetupWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create controller", "controller", "CustomInfoKey")
+		os.Exit(1)
+	}
+
 	if err := (&webhook.OrganizationValidator{}).SetupWebhookWithManager(mgr); err != nil {
 		setupLog.Error(err, "unable to create webhook", "webhook", "OrganizationValidator")
 		os.Exit(1)
@@ -229,6 +237,11 @@ func main() {
 
 	if err := (&webhook.AutoinstallDistributionValidator{}).SetupWebhookWithManager(mgr); err != nil {
 		setupLog.Error(err, "unable to create webhook", "webhook", "AutoinstallDistributionValidator")
+		os.Exit(1)
+	}
+
+	if err := (&webhook.CustomInfoKeyValidator{}).SetupWebhookWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create webhook", "webhook", "CustomInfoKeyValidator")
 		os.Exit(1)
 	}