+++ internal/controller/constants.go
@@ -45,4 +45,5 @@ const (
 	adFinalizer      = uyuniv1.Group + "/autoinstalldistribution"
 	apFinalizer      = uyuniv1.Group + "/autoinstallprofile"
 	ibFinalizer      = uyuniv1.Group + "/imagebuild"
+	cikFinalizer     = uyuniv1.Group + "/custominfokey"
 )